Remove old taxonomy model
