+++ bims/models/__init__.py
@@ -1,7 +1,6 @@
 from bims.models.location_type import *  # noqa
 from bims.models.location_site import *  # noqa
 from bims.models.iucn_status import *  # noqa
-from bims.models.taxon import *  # noqa
 from bims.models.survey import *  # noqa
 from bims.models.biological_collection_record import *  # noqa
 from bims.models.profile import Profile
