+++ bims/tasks/email_admins.py
@@ -9,13 +9,14 @@ def email_admins():
     from django.template.loader import get_template
     from django.utils.timezone import now
     from bims.conf import TRACK_PAGEVIEWS
-    from bims.models import Taxon, BiologicalCollectionRecord, Visitor, \
-        Pageview
+    from bims.models import (
+        BiologicalCollectionRecord, Visitor, Pageview, Taxonomy
+    )
 
     date_to = now()
     date_from = date_to - timedelta(days=7)
 
-    taxon_count = Taxon.objects.count()
+    taxon_count = Taxonomy.objects.count()
     collections_count = \
         BiologicalCollectionRecord.objects.filter(validated=True).count()
