+++ bims/templates/documents/document_metadata.html
@@ -106,46 +106,7 @@
 
             var url = document.URL.split('/');
             var docid = url[url.length-2];
-
-            get_chosen_taxon($taxonLinkSelect, docid);
         });
-
-        function get_chosen_taxon(div, docid) {
-            $.get('/api/list-taxon-for-document/'+docid+'/',
-                function (data) {
-                    var selected = [];
-                    if (data.length > 0) {
-                        $.each(data, function (key, value) {
-                            var option = new Option(value.scientific_name, value.id);
-                            option.selected = true;
-                            div.append(option);
-                            selected.push(value.id);
-                        });
-                        div.trigger('change');
-                    }
-                    get_taxon(div, '/api/list-taxon/', selected);
-                })
-        }
-
-        function get_taxon(div, url, selected) {
-            $.get(url,
-                function (data) {
-                    if (!data.hasOwnProperty('results')) {
-                        return true;
-                    }
-                    $.each(data.results, function (key, value) {
-                        if ($.inArray(selected, value.id) === -1) {
-                            div.append("<option value='"+value.id+"'>"+value.scientific_name+"</option>");
-                        }
-                    });
-                    if (data.hasOwnProperty('next')) {
-                        if (data['next']) {
-                            get_taxon(div, data['next'], selected);
-                        }
-                    }
-                }
-            );
-        }
 
     </script>
     <style>
